requirements.txt
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==2.0.12
Faker>=13.12.0
idna==3.3
iniconfig==1.1.1
jsonschema==4.22.0
packaging==21.3
pluggy==1.5.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==8.3.3
python-dateutil==2.8.2
requests==2.32.0
six==1.16.0
tomli==2.0.1
urllib3==2.2.2
pytest-repeat==0.9.1
pytest-asyncio>=0.23.7
pydantic==2.7.1
pytest-rerunfailures~=14.0
ruff==0.4.2
allure-pytest>=2.8.18
allure-python-commons~=2.13.5
httpx~=0.27.2
swagger-parser>=1.0.2
matplotlib>=3.9.2